st_s3kv_connection/requirements.txt
pycryptodome
cloudpickle
boto3

